primary entity abstraction
