+++ labstep/protocol.py
@@ -1,11 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+from .primaryEntity import PrimaryEntity
 from .entity import Entity, getEntity, getEntities, newEntity, editEntity
 from .helpers import (listToClass, getTime, createdAtFrom, createdAtTo)
-from .comment import addCommentWithFile, getComments
-from .tag import tag, getAttachedTags
-from .permissions import getPermissions, newPermission, transferOwnership
 
 
 def getProtocol(user, protocol_id):
@@ -252,7 +250,7 @@ class ProtocolTimer(Entity):
         return editEntity(self, fields)
 
 
-class Protocol(Entity):
+class Protocol(PrimaryEntity):
     """
     Represents a Protocol on Labstep.
 
@@ -302,96 +300,6 @@ class Protocol(Entity):
             my_protocol.delete()
         """
         return editProtocol(self, deleted_at=getTime())
-
-    def addComment(self, body, filepath=None):
-        """
-        Add a comment and/or file to a Labstep Protocol.
-
-        Parameters
-        ----------
-        body (str)
-            The body of the comment.
-        filepath (str)
-            A Labstep File entity to attach to the comment,
-            including the filepath.
-
-        Returns
-        -------
-        :class:`~labstep.comment.Comment`
-            The comment added.
-
-        Example
-        -------
-        ::
-
-            my_protocol = user.getProtocol(17000)
-            my_protocol.addComment(body='I am commenting!',
-                                   filepath='pwd/file_to_upload.dat')
-        """
-        return addCommentWithFile(self, body, filepath)
-
-    def getComments(self, count=100):
-        """
-        Retrieve the Comments attached to this Labstep Entity.
-
-        Returns
-        -------
-        List[:class:`~labstep.comment.Comment`]
-            List of the comments attached.
-
-        Example
-        -------
-        ::
-
-            entity = user.getProtocol(17000)
-            comments = entity.getComments()
-            comments[0].attributes()
-        """
-        return getComments(self, count)
-
-    def addTag(self, name):
-        """
-        Add a tag to the Protocol (creates a
-        new tag if none exists).
-
-        Parameters
-        ----------
-        name (str)
-            The name of the tag to create.
-
-        Returns
-        -------
-        :class:`~labstep.protocol.Protocol`
-            The Protocol that was tagged.
-
-        Example
-        -------
-        ::
-
-            my_protocol = user.getProtocol(17000)
-            my_protocol.addTag(name='My Tag')
-        """
-        tag(self, name)
-        return self
-
-    def getTags(self):
-        """
-        Retrieve the Tags attached to a this Labstep Entity.
-
-        Returns
-        -------
-        List[:class:`~labstep.tag.Tag`]
-            List of the tags attached.
-
-        Example
-        -------
-        ::
-
-            entity = user.getProtocol(17000)
-            tags = entity.getTags()
-            tags[0].attributes()
-        """
-        return getAttachedTags(self)
 
     def getSteps(self):
         """
@@ -595,42 +503,3 @@ class Protocol(Entity):
         """
         tables = self.last_version['protocol_tables']
         return listToClass(tables, ProtocolTable, self.__user__)
-
-    def getPermissions(self):
-        """
-        Returns the sharing permissions for the Protocol.
-
-        Returns
-        -------
-        List[:class:`~labstep.permissions.Permission`]
-        """
-        return getPermissions(self)
-
-    def shareWith(self, workspace_id, permission='view'):
-        """
-        Shares the Protocol with another Workspace.
-
-        Parameters
-        ----------
-        workspace_id (int)
-            The id of the workspace to share with
-
-        permission (str)
-            Permission to share with. Can be 'view' or 'edit'
-
-        Returns
-        -------
-        None
-        """
-        return newPermission(self, workspace_id, permission)
-
-    def transferOwnership(self, workspace_id):
-        """
-        Transfer ownership of the Entity to a different Workspace
-
-        Parameters
-        ----------
-        workspace_id (int)
-            The id of the workspace to transfer ownership to
-        """
-        return transferOwnership(self, workspace_id)
